remove `Header.svelte` in favor of a generic workspace header component that can be configured for multiple different use cases
