+++ src/routes/_surfaces/workspace/ModelWorkspaceHeader.svelte
@@ -1,0 +1,46 @@
+<script lang="ts">
+  import { getContext } from "svelte";
+  import {
+    ApplicationStore,
+    dataModelerService,
+  } from "$lib/application-state-stores/application-store";
+
+  import type { PersistentModelStore } from "$lib/application-state-stores/model-stores";
+  import type { PersistentModelEntity } from "$common/data-modeler-state-service/entity-state-service/PersistentModelEntityService";
+  import { ActionStatus } from "$common/data-modeler-service/response/ActionResponse";
+  import WorkspaceHeader from "./WorkspaceHeader.svelte";
+
+  const store = getContext("rill:app:store") as ApplicationStore;
+  const persistentModelStore = getContext(
+    "rill:app:persistent-model-store"
+  ) as PersistentModelStore;
+
+  function formatModelName(str) {
+    let output = str.trim().replaceAll(" ", "_").replace(/\.sql/, "");
+    return output;
+  }
+
+  let currentModel: PersistentModelEntity;
+  $: if ($store?.activeEntity && $persistentModelStore?.entities)
+    currentModel = $persistentModelStore.entities.find(
+      (q) => q.id === $store.activeEntity.id
+    );
+
+  let titleInput = currentModel?.name;
+  $: titleInput = currentModel?.name;
+
+  // FIXME: this should eventually be a redux action dispatcher `onChangeAction`
+  const onChangeCallback = async (e) => {
+    if (currentModel?.id) {
+      const resp = await dataModelerService.dispatch("updateModelName", [
+        currentModel?.id,
+        formatModelName(e.target.value),
+      ]);
+      if (resp.status === ActionStatus.Failure) {
+        e.target.value = currentModel.name;
+      }
+    }
+  };
+</script>
+
+<WorkspaceHeader {...{ titleInput, onChangeCallback }} />
